+++ packages/client/src/rtc/__tests__/Publisher.test.ts
@@ -17,7 +17,8 @@ vi.mock('../../StreamSfuClient', () => {
   };
 });
 
-vi.mock('../codecs', () => {
+vi.mock('../codecs', async () => {
+  const codecs = await vi.importActual('../codecs');
   return {
     getPreferredCodecs: vi.fn((): RTCRtpCodecCapability[] => [
       {
@@ -27,6 +28,8 @@ vi.mock('../codecs', () => {
         sdpFmtpLine: 'profile-level-id=42e01f',
       },
     ]),
+    getOptimalVideoCodec: codecs.getOptimalVideoCodec,
+    isSvcCodec: codecs.isSvcCodec,
   };
 });
 
@@ -136,11 +139,7 @@ describe('Publisher', () => {
     vi.spyOn(transceiver.sender, 'track', 'get').mockReturnValue(newTrack);
 
     expect(track.stop).toHaveBeenCalled();
-    expect(track.removeEventListener).toHaveBeenCalledWith(
-      'ended',
-      expect.any(Function),
-    );
-    expect(newTrack.addEventListener).toHaveBeenCalledWith(
+    expect(newTrack.addEventListener).not.toHaveBeenCalledWith(
       'ended',
       expect.any(Function),
     );
@@ -154,7 +153,7 @@ describe('Publisher', () => {
     );
   });
 
-  it('can publish and un-pubish with just enabling and disabling tracks', async () => {
+  it('can publish and un-publish with just enabling and disabling tracks', async () => {
     const mediaStream = new MediaStream();
     const track = new MediaStreamTrack();
     mediaStream.addTrack(track);
@@ -275,6 +274,196 @@ describe('Publisher', () => {
       publisher['pc'].iceConnectionState = 'disconnected';
       publisher['onIceConnectionStateChange']();
       expect(publisher.restartIce).toHaveBeenCalled();
+    });
+  });
+
+  describe('changePublishQuality', () => {
+    it('can dynamically activate/deactivate simulcast layers', async () => {
+      const transceiver = new RTCRtpTransceiver();
+      const setParametersSpy = vi
+        .spyOn(transceiver.sender, 'setParameters')
+        .mockResolvedValue();
+      const getParametersSpy = vi
+        .spyOn(transceiver.sender, 'getParameters')
+        .mockReturnValue({
+          codecs: [
+            // @ts-expect-error incomplete data
+            { mimeType: 'video/VP8' },
+            // @ts-expect-error incomplete data
+            { mimeType: 'video/VP9' },
+            // @ts-expect-error incomplete data
+            { mimeType: 'video/H264' },
+            // @ts-expect-error incomplete data
+            { mimeType: 'video/AV1' },
+          ],
+          encodings: [
+            { rid: 'q', active: true },
+            { rid: 'h', active: true },
+            { rid: 'f', active: true },
+          ],
+        });
+
+      // inject the transceiver
+      publisher['transceiverCache'].set(TrackType.VIDEO, transceiver);
+
+      await publisher['changePublishQuality']([
+        {
+          name: 'q',
+          active: true,
+          maxBitrate: 100,
+          scaleResolutionDownBy: 4,
+          maxFramerate: 30,
+          scalabilityMode: '',
+        },
+        {
+          name: 'h',
+          active: false,
+          maxBitrate: 150,
+          scaleResolutionDownBy: 2,
+          maxFramerate: 30,
+          scalabilityMode: '',
+        },
+        {
+          name: 'f',
+          active: true,
+          maxBitrate: 200,
+          scaleResolutionDownBy: 1,
+          maxFramerate: 30,
+          scalabilityMode: '',
+        },
+      ]);
+
+      expect(getParametersSpy).toHaveBeenCalled();
+      expect(setParametersSpy).toHaveBeenCalled();
+      expect(setParametersSpy.mock.calls[0][0].encodings).toEqual([
+        {
+          rid: 'q',
+          active: true,
+          maxBitrate: 100,
+          scaleResolutionDownBy: 4,
+          maxFramerate: 30,
+        },
+        {
+          rid: 'h',
+          active: false,
+          maxBitrate: 150,
+          scaleResolutionDownBy: 2,
+          maxFramerate: 30,
+        },
+        {
+          rid: 'f',
+          active: true,
+          maxBitrate: 200,
+          scaleResolutionDownBy: 1,
+          maxFramerate: 30,
+        },
+      ]);
+    });
+
+    it('can dynamically update scalability mode in SVC', async () => {
+      const transceiver = new RTCRtpTransceiver();
+      const setParametersSpy = vi
+        .spyOn(transceiver.sender, 'setParameters')
+        .mockResolvedValue();
+      const getParametersSpy = vi
+        .spyOn(transceiver.sender, 'getParameters')
+        .mockReturnValue({
+          codecs: [
+            // @ts-expect-error incomplete data
+            { mimeType: 'video/VP9' },
+            // @ts-expect-error incomplete data
+            { mimeType: 'video/AV1' },
+            // @ts-expect-error incomplete data
+            { mimeType: 'video/VP8' },
+            // @ts-expect-error incomplete data
+            { mimeType: 'video/H264' },
+          ],
+          encodings: [
+            {
+              rid: 'q',
+              active: true,
+              maxBitrate: 100,
+              // @ts-expect-error not in the standard lib yet
+              scalabilityMode: 'L3T3_KEY',
+            },
+          ],
+        });
+
+      // inject the transceiver
+      publisher['transceiverCache'].set(TrackType.VIDEO, transceiver);
+
+      await publisher['changePublishQuality']([
+        {
+          name: 'q',
+          active: true,
+          maxBitrate: 50,
+          scaleResolutionDownBy: 1,
+          maxFramerate: 30,
+          scalabilityMode: 'L1T3',
+        },
+      ]);
+
+      expect(getParametersSpy).toHaveBeenCalled();
+      expect(setParametersSpy).toHaveBeenCalled();
+      expect(setParametersSpy.mock.calls[0][0].encodings).toEqual([
+        {
+          rid: 'q',
+          active: true,
+          maxBitrate: 50,
+          scaleResolutionDownBy: 1,
+          maxFramerate: 30,
+          scalabilityMode: 'L1T3',
+        },
+      ]);
+    });
+
+    it('supports empty rid in SVC', async () => {
+      const transceiver = new RTCRtpTransceiver();
+      const setParametersSpy = vi
+        .spyOn(transceiver.sender, 'setParameters')
+        .mockResolvedValue();
+      const getParametersSpy = vi
+        .spyOn(transceiver.sender, 'getParameters')
+        .mockReturnValue({
+          codecs: [
+            // @ts-expect-error incomplete data
+            { mimeType: 'video/VP9' },
+          ],
+          encodings: [
+            {
+              rid: undefined, // empty rid
+              active: true,
+              // @ts-expect-error not in the standard lib yet
+              scalabilityMode: 'L3T3_KEY',
+            },
+          ],
+        });
+
+      // inject the transceiver
+      publisher['transceiverCache'].set(TrackType.VIDEO, transceiver);
+
+      await publisher['changePublishQuality']([
+        {
+          name: 'q',
+          active: true,
+          maxBitrate: 50,
+          scaleResolutionDownBy: 1,
+          maxFramerate: 30,
+          scalabilityMode: 'L1T3',
+        },
+      ]);
+
+      expect(getParametersSpy).toHaveBeenCalled();
+      expect(setParametersSpy).toHaveBeenCalled();
+      expect(setParametersSpy.mock.calls[0][0].encodings).toEqual([
+        {
+          active: true,
+          maxBitrate: 50,
+          scaleResolutionDownBy: 1,
+          maxFramerate: 30,
+          scalabilityMode: 'L1T3',
+        },
+      ]);
     });
   });
 });