+++ sample-apps/react-native/dogfood/src/components/ParticipantsInfoList.tsx
@@ -156,6 +156,7 @@ export const ParticipantsInfoList = ({
           transparent
           visible={!!selectedParticipant}
           onRequestClose={() => setSelectedParticipant(undefined)}
+          supportedOrientations={['portrait', 'landscape']}
         >
           <>
             {/*independent background, needed due to desired opacity only